Rename files
These are not just about recovery keys.
